Fixed component file names case.
