+++ app/components/ServerView/ServerKeyTree/ServerListItemIcon.js
